md uses new load model method
